Separate pdf parsing logic
